Fix BFloat16 numpy crashes on Gemma 4 + smarter patcher
7 BLME tasks crashed on bf16 models with 'Got unsupported ScalarType
BFloat16' because numpy can't convert bfloat16 directly. Added .float()
casts before .numpy() / .cpu().numpy() in:
- geometry/correlation_dimension.py:88
- geometry/positional_decay.py:70
- interpretability/attention.py:72
- interpretability/sparsity.py:54
- interpretability/superposition.py:126
- interpretability/attention_graph.py:115
- representation_engineering.py:175

This fixes 28 missing cells (7 tasks × 4 Gemma 4 models) and unblocks
similar bf16-only models (Mixtral, future Llama variants, etc.).

scripts/patch_failed_tasks.py improvements:
- Compute per-model patch list inside _patch_model — only re-runs the
  tasks that actually need fixing for each individual model, even when
  a shared task list is passed at the CLI. Avoids wasted re-runs of
  tasks that already succeeded.
- New --task-timeout flag (default 600) propagated to blme.cli evaluate.
  Bumping to 1800 unblocks the slow gemma4 tasks (geometry_collapse,
  geometry_hubness, interpretability_waa) that timed out at the default.
- New --subprocess-timeout flag (default 7200) hard-caps the wall clock
  per model.
- patched_tasks field now accumulates across multiple patch passes
  instead of overwriting.

+++ src/blme/tasks/geometry/positional_decay.py
@@ -67,7 +67,8 @@ class PositionalAttentionDecayTask(DiagnosticTask):
                     if attn_entry is None:
                         continue
                     attn_matrix = attn_entry[0]  # (num_heads, seq_len, seq_len)
-                    mean_attn = attn_matrix.mean(dim=0).cpu().numpy()  # (seq_len, seq_len)
+                    # .float() so bf16 models (Gemma 4 etc.) don't crash on .numpy()
+                    mean_attn = attn_matrix.mean(dim=0).float().cpu().numpy()  # (seq_len, seq_len)
 
                     distances = []
                     attentions = []
